+++ src/gui/CMakeLists.txt
@@ -7,6 +7,7 @@ add_executable(zappy_gui
         App.cpp
         App+Listener.cpp
         Button.cpp
+        Mediator.cpp
         )
 
 find_package(OGRE QUIET)
