+++ src/gui/Observer.hpp
@@ -10,6 +10,8 @@
 
 #include <string>
 namespace Zappy::GUI {
+    class Mediator;
+    class ServerData;
     /**
      * @brief Interface for the Observer pattern
      * @details This interface is used to implement the Observer pattern, it is representing a class that can be
@@ -18,7 +20,8 @@ namespace Zappy::GUI {
     class Observer
     {
         public:
-            virtual ~Observer() = default;
+            explicit Observer(Mediator &mediator, ServerData &serverData);
+            virtual ~Observer();
 
             /**
              * @brief Method called by the Publisher to notify the Subscriber
@@ -26,6 +29,17 @@ namespace Zappy::GUI {
              * @param aNotification The notification sent by the Publisher
              */
             virtual void getNotified(const std::string &aNotification) = 0;
+
+            /**
+             * @brief Alert the observer of an object
+             *
+             * @param aNotification the notification
+             */
+            void sendAlert(const std::string &aNotification);
+
+        protected:
+            Mediator &_mediator;
+            ServerData &_serverData;
     };
 } // namespace Zappy::GUI
