+++ src/gui/cli/ServerData.hpp
@@ -3,6 +3,8 @@
 #include <iostream>
 #include <vector>
 #include "EggData.hpp"
+#include "Mediator.hpp"
+#include "Observer.hpp"
 #include "PlayerData.hpp"
 
 namespace Zappy::GUI {
@@ -18,11 +20,6 @@ namespace Zappy::GUI {
 
     struct ServerData
     {
-            static ServerData &getInstance()
-            {
-                static ServerData instance;
-                return instance;
-            }
             std::pair<unsigned int, unsigned int> _mapSize;
             std::vector<TileContent> _mapTiles;
             std::vector<std::string> _teamNames;
