+++ src/gui/main.cpp
@@ -8,6 +8,7 @@
 #include <thread>
 #include "App.hpp"
 #include "ClientApi.hpp"
+#include "Mediator.hpp"
 #include "ParserData.hpp"
 
 int main(int argc, char **argv)
@@ -15,13 +16,9 @@ int main(int argc, char **argv)
     try {
         Zappy::GUI::ParserData parserData("127.0.0.1", 0, "");
         parserData.parseData(argc, argv);
-        Zappy::GUI::ClientApi myClientApi(parserData.getAddress(), parserData.getPort(), "GRAPHIC");
-        myClientApi.joinGame();
-        std::thread myCliThread(&Zappy::GUI::ClientApi::run, &myClientApi);
-        Zappy::GUI::App ctx(myClientApi, "Zappy");
-        // map size | map content | teams | time unit | player connected
-        ctx.getRoot()->startRendering();
-        myCliThread.join();
+        Zappy::GUI::Mediator mediator(parserData);
+
+        mediator.start();
     } catch (const std::exception &e) {
         std::cerr << e.what() << std::endl;
     }
